+++ harvest.lua
@@ -6,7 +6,7 @@
 local loopCount = -1
 
 -- How long to wait between harvesting (in seconds)
-local sleepSecs = 900
+local sleepSecs = 90
 
 -- Starting position:
 -- bottom right: startPos = "right"
@@ -68,11 +68,22 @@ local function digForward()
     end
 end
 local function dropAll()
+    local moveExtra = turtle.forward()
     for slot = 1, 9 do
-        turtle.select(slot)
-        turtle.drop()
+        if turtle.getItemCount(slot) > 0 then
+            turtle.select(slot)
+            turtle.drop()
+            sleep(0.5)
+        end
     end
     turtle.select(1)
+    if moveExtra then
+        if not turtle.back() then
+            aboutFace()
+            digForward()
+            aboutFace()
+        end
+    end
 end
 
 local skipIndex = 0
